requirements.txt
streamlit
python-dotenv
langchain_community
pymongo
motor
pydantic
cryptography
youtube-transcript-api
pytube